dom/tests/mochitest/gamepad/test_check_timestamp.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test Gamepad.timestamp</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript" src="mock_gamepad.js"></script>
<script class="testbody" type="text/javascript">

window.addEventListener("gamepadbuttondown", buttonpresshandler);

var index;
var timea=0;
var firstPress = true;
var testOver = false;

SimpleTest.waitForExplicitFinish();
runGamepadTest(checkTimestamp);

function checkTimestamp(){
  GamepadService.addGamepad("test gamepad 1",
                            GamepadService.standardMapping,
                            4,
                            2).then(function(i) {
                              index = i;
                              // Press a button to make the gamepad visible
                              // to the page.
                              GamepadService.newButtonEvent(index, 0, true);
                              GamepadService.newButtonEvent(index, 0, true);
                              ok(true, "test");
                            });
}

function cleanup(){
  SpecialPowers.executeSoon(function() {
    GamepadService.removeGamepad(index);
    SimpleTest.finish();
  });
}

function buttonpresshandler(e) {
  if (testOver) {
    return;
  }
  if (timea == 0){
    timea = e.gamepad.timestamp;
  } else {
    ok(timea <= e.gamepad.timestamp, "Timestamp less than last timestamp");
  }
  GamepadService.newButtonEvent(index, 0, false);
  if (!firstPress) {
    testOver = true;
    SpecialPowers.executeSoon(cleanup);
  } else {
    firstPress = false;
  }
}

</script>
</body>
</html>
